requirements.txt
django-cors-headers=0.12